requirements.txt
# Backend Dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==12.0
redis==5.0.1
pandas==2.1.3
numpy>=1.26.0
requests==2.31.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
aiofiles==23.2.1
httpx==0.25.2
asyncio-mqtt==0.16.1

# Data Analysis & ML
scikit-learn==1.3.2
ta==0.10.2
yfinance==0.2.28
beautifulsoup4==4.12.2
textblob==0.17.1
nltk==3.8.1
torch>=2.0.0
xgboost>=1.7.0
lightgbm>=4.0.0
feedparser>=6.0.10
spacy>=3.7.0
pymc>=5.0.0
arviz>=0.15.0
theano-pymc>=1.4.0

# Database
sqlalchemy==2.0.23
alembic==1.13.0
psycopg2-binary==2.9.9

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Development
black==23.11.0
flake8==6.1.0
mypy==1.7.1